requirements.txt
flask
gunicorn
requests
cairocffi
CairoSVG
pillow
svgwrite
tinycss2
lxml
python-dotenv
flask-limiter
